A bugfix for Crop module
As I have mentioned in Issue #19, `options.format` isn't defined initially. Hence it always gets assigned the value `png`.
This doesn't go well with JPEG images. So I changed it to `jpg`. This does the trick temporarily but then alpha images aren't supported. So could this be implemented till we discuss #19?